requirements.txt
matplotlib
numpy
pandas
pydash
wget
opencv-python
pillow
imutils
scikit-image
librosa
seaborn
tqdm
xgboost
PyWavelets
tensorflow-macos
tensorflow-metal
Keras
h5py